amd64: enable i386-only drivers for x86_64
Widens the per-category SUBDIR gates so all previously i386-only
drivers also build for amd64, and applies the minimal source fixes
needed to clear -Werror on 64-bit (the same patterns we've been
applying since at_wini):

- (uintptr_t) cast on PCI BAR / port values that are uint32_t but
  fed to vm_map_phys() or volatile-pointer dereferences expecting
  a host pointer (3c90x, atl2, dp8390, e1000, lance, ahci, ti1225,
  cs4281)
- %d -> %zu / %lu for printf of size_t / unsigned long arguments
  (fxp, lance, fbd, filter, amddev)
- fbd's rule_post_hook(int *) takes an int; pass through a local
  int instead of &(ssize_t) so the call stays well-typed

Categories migrated:
  audio    als4000 cmi8738 cs4281 es1370 es1371 sb16 trident
  bus      ti1225
  iommu    amddev
  net      3c90x atl2 dec21140A dp8390 dpeth e1000 fxp ip1000
           lance rtl8139 rtl8169 virtio_net vt6105
  printer  printer
  storage  ahci fbd filter
  vmm_guest vbox

floppy is intentionally left i386-only.  igc and at_wini stay
gated for amd64 as before.  All 25 newly-enabled drivers compile
and link cleanly with the amd64 cross-toolchain; ramdisk/set-list
packaging and runtime smoke tests are not part of this commit.

+++ minix/drivers/net/lance/lance.c
@@ -758,10 +758,10 @@ static void lance_init_hw(ether_card_t *ec, netdriver_addr_t *addr,
    /* Allocate memory */
    if ((lance_buf = alloc_contig(LANCE_BUF_SIZE, AC_ALIGN4K|AC_LOWER16M,
      &lance_buf_phys)) == NULL)
-      panic("alloc_contig failed: %d", LANCE_BUF_SIZE);
+      panic("alloc_contig failed: %lu", (unsigned long)LANCE_BUF_SIZE);
 
    l = (vir_bytes)lance_buf;
-   lp = (struct lance_interface *)l;
+   lp = (struct lance_interface *)(uintptr_t)l;
 
    /* disable Tx and Rx */
    lp->init_block.mode = LANCE_CSR15_DTX|LANCE_CSR15_DRX;
